# Allow slashes in unquoted attribute values