--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,8 @@
 node_modules/
 npm-debug.log
 .DS_Store
+tmp/
+*.log
+.eslintcache
+*.tgz
+package-lock.json