--- a/.travis.yml
+++ b/.travis.yml
@@ -1,7 +1,0 @@
-language: node_js
-node_js:
-  - "node"
-os:
-  - linux
-  - osx
-  - windows