--- a/src/HtmlParser.js
+++ b/src/HtmlParser.js
@@ -18,7 +18,7 @@ class SeqMatcher {
 }
 
 module.exports = class HtmlParser extends Transform {
-  constructor({preserveWS} = {preserveWS: false}){
+  constructor({preserveWS} = {}){
     super({readableObjectMode: true})
     this.preserveWS = preserveWS;
     this.endScript = new SeqMatcher("</script>")
@@ -36,27 +36,41 @@ module.exports = class HtmlParser extends Transform {
     this.text = [];
     this.data = {};
 
-    this.flags = {
-      isClose: null,
-      isSelfClose: null,
-      hasEqual: null,
-      valStartChar: null,
-    };
+    this.isClose = null;
+    this.isSelfClose = null;
+    this.hasEqual = null;
+    this.valStartChar = null;
+  }
+  reset() {
+    this.endScript.reset()
+    this.endStyle.reset()
+    this.beginComment.reset()
+    this.endComment.reset()
+
+    this.curPos = 0;
+    this.minPos = 0;
+    this.state = TEXT;
+
+    this.cache = "";
+    this.name = "";
+    this.key = "";
+    this.text = [];
+    this.data = {};
+
+    this.isClose = null;
+    this.isSelfClose = null;
+    this.hasEqual = null;
+    this.valStartChar = null;
   }
   _transform(chunk, encoding, done){
     if (chunk === null) return this.end();
-    this.parse(this.addToCache(chunk));
-    done(null)
-  }
-  _flush(done){
-    this.flushAllText();
-    this.reset();
-    done(null)
-  }
-  parse(cacheLen) {
+
+    const cache = this.cache += chunk;
+    const cacheLen = cache.length
+
     let i = this.curPos, v = this.minPos, s = this.state, c
-    while (i < cacheLen){
-      c = this.cache.charCodeAt(i)
+    while (i < cacheLen) {
+      c = cache.charCodeAt(i)
       if (s === TEXT){
         if (!this.preserveWS && (c === 32 || c >= 9 && c <= 13)) // ws
           v < i && this.text.push(this.getCache(v, i)), v = i + 1
@@ -65,19 +79,19 @@ module.exports = class HtmlParser extends Transform {
       } else if (s === NODE){
         if (c === 62) // >
           this.key && this.flushKey(), s = this.flushNode(), v = i + 1
-        else if (c === 47 && !this.flags.hasEqual) // /
-          this.flags.isClose = !(this.flags.isSelfClose = !!this.name)
+        else if (c === 47 && !this.hasEqual) // /
+          this.isClose = !(this.isSelfClose = !!this.name)
         else if (c !== 32 && (c < 9 || c > 13)){ // !ws
           if (!this.name) // name start
             this.beginComment.found(c), v = i, s = NAME
           else if (!this.key) // key start
             v = i, s = KEY
           else if (c === 61) // =
-            this.flags.hasEqual = true
-          else if (!this.flags.hasEqual) // next key
+            this.hasEqual = true
+          else if (!this.hasEqual) // next key
             this.flushKey(), v = i, s = KEY
           else if (c === 34 || c === 39) // ', "
-            v = i + 1, this.flags.valStartChar = c, s = VALUE
+            v = i + 1, this.valStartChar = c, s = VALUE
           else // un-quoted val
             v = i, s = VALUE
         }
@@ -87,21 +101,21 @@ module.exports = class HtmlParser extends Transform {
         else if (c === 32 || c >= 9 && c <= 13) // ws
           this.name = this.getCache(v, i), s = NODE, v = i + 1
         else if (c === 47) // /
-          this.flags.isSelfClose = true, this.name = this.getCache(v, i), s = NODE, v = i + 1
+          this.isSelfClose = true, this.name = this.getCache(v, i), s = NODE, v = i + 1
         else if (c === 62) // >
           this.name = this.getCache(v, i), s = this.flushNode(), v = i + 1
       } else if (s === KEY){
         if (c === 32 || c >= 9 && c <= 13) // ws
           this.key = this.getCache(v, i), s = NODE, v = i + 1
         else if (c === 61) // =
-          this.flags.hasEqual = true, this.key = this.getCache(v, i), s = NODE, v = i + 1
+          this.hasEqual = true, this.key = this.getCache(v, i), s = NODE, v = i + 1
         else if (c === 47) // /
-          this.flags.isSelfClose = true, this.key = this.getCache(v, i), s = NODE, v = i + 1
+          this.isSelfClose = true, this.key = this.getCache(v, i), s = NODE, v = i + 1
         else if (c === 62) // >
           this.flushKey(v,i), s = this.flushNode(), v = i + 1
       } else if (s === VALUE){
-        if (this.flags.valStartChar != null){
-          if (c === this.flags.valStartChar) // found end quote
+        if (this.valStartChar != null){
+          if (c === this.valStartChar) // found end quote
             this.flushVal(v,i), s = NODE, v = i + 1
         } else if (c === 32 || c >= 9 && c <= 13) // ws
           this.flushVal(v,i), s = NODE, v = i + 1
@@ -113,74 +127,53 @@ module.exports = class HtmlParser extends Transform {
         s = this.flushSpecialNode(v, i-8, "script"), v = i + 1
       else if (s === STYLE && this.endStyle.found(c))
         s = this.flushSpecialNode(v, i-7, "style"), v = i + 1
-      i = i + 1
+      i++
     }
-    this.cache = this.cache.substr(v);
+
+    this.cache = this.cache.substring(v);
     this.curPos = i - v;
     this.minPos = 0;
     this.state = s;
+
+    done(null)
   }
-  reset() {
-    this.endStyle.reset();
-    this.endScript.reset();
-    this.endComment.reset();
-    this.beginComment.reset();
-
-    this.state = TEXT;
-    this.cache = "";
-    this.name = "";
-    this.key = "";
-    this.text = [];
-    this.data = {};
-
-    this.curPos = 0;
-    this.minPos = 0;
-
-    this.flags = {
-      isClose: null,
-      isSelfClose: null,
-      hasEqual: null,
-      valStartChar: null,
-    };
+  _flush(done){
+    this.flushText(this.minPos, this.curPos);
+    this.reset();
+    done(null)
   }
   getCache(start, end) {
-    return this.cache.substr(start, end-start);
-  }
-  addToCache(chunk) {
-    return (this.cache += chunk).length;
+    return this.cache.substring(start, end);
   }
   flushKey(v, i) {
-    return (this.key = this.data[this.key || this.cache.substr(v, i-v)] = "");
+    return (this.key = this.data[this.key || this.cache.substring(v, i)] = "");
   }
   flushVal(v, i) {
-    return (this.data[this.key] = this.cache.substr(v, i-v), this.key = "", this.flags.valStartChar = this.flags.hasEqual = null);
+    return (this.data[this.key] = this.cache.substring(v, i), this.key = "", this.valStartChar = this.hasEqual = null);
   }
   flushNode() {
     const name = this.name;
-    if (!this.flags.isClose) this.push({name, data: this.data})
-    if (this.flags.isSelfClose || this.flags.isClose) this.push({name})
+    if (!this.isClose) this.push({name, data: this.data})
+    if (this.isSelfClose || this.isClose) this.push({name})
     const s = name === "script" ? SCRIPT : name === "style" ? STYLE : name === "!--" ? COMMENT : TEXT
     this.data = {};
     this.name = "";
-    this.flags.isClose = this.flags.isSelfClose = null
+    this.isClose = this.isSelfClose = null
     return s
   }
   flushSpecialNode(v, i, name) {
-    const text = this.cache.substr(v, i-v)
+    const text = this.cache.substring(v, i)
     text && this.push({text}), this.push({name})
     return TEXT
   }
   flushText(v, i) {
     if (v < i) {
-      this.text.push(this.cache.substr(v, i-v))
+      this.text.push(this.cache.substring(v, i))
       this.push({text: this.text.join(" ")})
-      this.text = []
+      this.text.length = 0;
     } else if (this.text.length) {
       this.push({text: this.text.join(" ")})
-      this.text = []
+      this.text.length = 0;
     }
-  }
-  flushAllText() {
-    return this.flushText(this.minPos, this.curPos);
   }
 }