Hello!

### Mad props!

Thank you so much for this fantastic HTML-parser! Great stuff!

We use this, amongst other things, to parse and process our sites output markup on request, so we really appreciate the speed of this thing.

### Issue

Unquoted attribute values containing `/` characters aren't parsed correctly.

Consider the following markup:

```xml
<a href=https://www.mozilla.com/> Absolute URL </a>
<a href=//example.com/> Scheme-relative URL </a>
<a href=/en-US/docs/Web/HTML/> Origin-relative URL </a>
<a href=./p/> Directory-relative URL </a>
```

Had there been quotes around the values we would get these lovely data chunks. Hooray!

```js
{ name: 'a', data: { href: 'https://www.mozilla.com/' } }
{ text: 'Absolute URL' }
{ name: 'a' }
{ name: 'a', data: { href: '//example.com/' } }
{ text: 'Scheme-relative URL' }
{ name: 'a' }
{ name: 'a', data: { href: '/en-US/docs/Web/HTML/' } }
{ text: 'Origin-relative URL' }
{ name: 'a' }
{ name: 'a', data: { href: './p/' } }
{ text: 'Directory-relative URL' }
{ name: 'a' }
```

Unfortunately we get this mess instead. Rats!

```js
{ name: 'a', data: { href: 'https:', 'www.mozilla.com': '' } }
{ name: 'a' }
{ text: 'Absolute URL' }
{ name: 'a' }
{ name: 'a', data: { href: 'example.com' } }
{ name: 'a' }
{ text: 'Scheme-relative URL' }
{ name: 'a' }
{ name: 'a', data: { href: 'en-US', docs: '', Web: '', HTML: '' } }
{ name: 'a' }
{ text: 'Origin-relative URL' }
{ name: 'a' }
{ name: 'a', data: { href: '.', p: '' } }
{ name: 'a' }
{ text: 'Directory-relative URL' }
{ name: 'a' }
```

### Fix

By bypassing / modifying a few conditions I was able to solve this.

I ran the speed tests and found no significant impact. Although the numbers don't match the ones in [the README](https://github.com/atlassubbed/atlas-html-stream#htmlparser-benchmark) at all (maybe I did it wrong?) they don't differ in between audits.

#### Before

- 177.659 ms/file ± 317.734
- 177.759 ms/file ± 316.590
- 180.044 ms/file ± 317.591

#### After

- 176.217 ms/file ± 316.235
- 176.564 ms/file ± 315.775
- 177.296 ms/file ± 317.872